// admin/src/api/generated/schemas/adminWalletDepositItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminWalletDepositItem = {
  id?: string;
  user_id?: string;
  asset_code?: string;
  chain_code?: string;
  deposit_address?: string;
  amount?: string;
  transaction_hash?: string;
  block_number?: string;
  confirmations?: number;
  status?: string;
  memo?: string;
  admin_id?: string;
  admin_note?: string;
  created_at?: string;
  updated_at?: string;
};